revert: unifying methods
