+++ packages/contracts/test/identityVAnchor/identityVAnchor.test.ts
@@ -82,7 +82,6 @@ describe('IdentityVAnchor for 2 max edges', () => {
   let alice: SignerWithAddress;
   let bob: SignerWithAddress;
   let carl: SignerWithAddress;
-  const dummyLeavesRecord: Record<string, Uint8Array[]> = {};
 
   const identity_vanchor_2_2_wasm_path = path.resolve(
     __dirname,
@@ -372,7 +371,6 @@ describe('IdentityVAnchor for 2 max edges', () => {
         alice.address,
         relayer,
         '',
-        dummyLeavesRecord,
         { keypair: aliceKeypair }
       );
 
@@ -412,7 +410,6 @@ describe('IdentityVAnchor for 2 max edges', () => {
         alice.address,
         relayer,
         '',
-        dummyLeavesRecord,
         { keypair: aliceKeypair }
       );
 
@@ -464,7 +461,6 @@ describe('IdentityVAnchor for 2 max edges', () => {
         alice.address,
         relayer,
         '',
-        dummyLeavesRecord,
         { keypair: aliceKeypair }
       )) as ContractReceipt;
 
@@ -513,7 +509,6 @@ describe('IdentityVAnchor for 2 max edges', () => {
         bob.address,
         relayer,
         '',
-        dummyLeavesRecord,
         { keypair: bobKeypair }
       );
 
@@ -542,7 +537,6 @@ describe('IdentityVAnchor for 2 max edges', () => {
           alice.address,
           relayer,
           '',
-          dummyLeavesRecord,
           { keypair: aliceKeypair }
         );
         aliceBalanceAfterDeposit = await token.balanceOf(alice.address);
@@ -573,7 +567,6 @@ describe('IdentityVAnchor for 2 max edges', () => {
           alice.address,
           relayer,
           '',
-          dummyLeavesRecord,
           { keypair: aliceKeypair }
         );
       });
@@ -593,7 +586,6 @@ describe('IdentityVAnchor for 2 max edges', () => {
           alice.address,
           relayer,
           '',
-          dummyLeavesRecord,
           { keypair: aliceKeypair }
         );
 
@@ -612,7 +604,6 @@ describe('IdentityVAnchor for 2 max edges', () => {
             alice.address,
             relayer,
             '',
-            dummyLeavesRecord,
             { keypair: aliceKeypair }
           )
         ).to.revertedWith('Input is already spent');
@@ -638,7 +629,6 @@ describe('IdentityVAnchor for 2 max edges', () => {
             alice.address,
             relayer,
             '',
-            dummyLeavesRecord,
             { keypair: aliceKeypair }
           )
         ).to.revertedWith('ERC20: transfer amount exceeds balance');
@@ -663,7 +653,6 @@ describe('IdentityVAnchor for 2 max edges', () => {
           alice.address,
           relayer,
           '',
-          dummyLeavesRecord,
           { keypair: aliceKeypair }
         );
         const aliceBalanceAfterSecondDeposit = await token.balanceOf(alice.address);
@@ -686,7 +675,6 @@ describe('IdentityVAnchor for 2 max edges', () => {
           alice.address,
           relayer,
           '',
-          dummyLeavesRecord,
           { keypair: aliceKeypair }
         );
         const aliceBalanceAfterJoin = await token.balanceOf(alice.address);
@@ -706,7 +694,6 @@ describe('IdentityVAnchor for 2 max edges', () => {
           alice.address,
           relayer,
           '',
-          dummyLeavesRecord,
           { keypair: aliceKeypair }
         );
 
@@ -735,7 +722,6 @@ describe('IdentityVAnchor for 2 max edges', () => {
           aliceETHAddress,
           relayer,
           '',
-          dummyLeavesRecord,
           { keypair: aliceKeypair }
         );
 
@@ -1452,7 +1438,6 @@ describe('IdentityVAnchor for 2 max edges', () => {
         '0',
         relayer,
         token.address,
-        dummyLeavesRecord,
         { keypair: aliceKeypair }
       );
       const balTokenAfterDepositSender = await token.balanceOf(alice.address);
@@ -1482,7 +1467,6 @@ describe('IdentityVAnchor for 2 max edges', () => {
         '0',
         relayer,
         token.address,
-        dummyLeavesRecord,
         { keypair: aliceKeypair }
       );
       const balTokenAfterDepositSender = await token.balanceOf(alice.address);
@@ -1519,7 +1503,6 @@ describe('IdentityVAnchor for 2 max edges', () => {
         alice.address,
         relayer,
         token.address,
-        dummyLeavesRecord,
         { keypair: aliceKeypair }
       );
